airflow/requirements.txt
pandas==2.2.3
numpy==2.0.1
psycopg2-binary==2.9.9